Refactor config helpers (getComponentConfigs and getTemplateAppComponentConfigs) to return a components config object. Move config file parsing logic into utils so it can be shared in all config helpers.
