+++ src/org/feedhenry/Utils.groovy
@@ -1,6 +1,8 @@
 #!/usr/bin/groovy
 package org.feedhenry
 
+import groovy.json.JsonSlurperClassic
+
 import java.text.SimpleDateFormat
 
 static def getReleaseBranch(version) {
@@ -48,6 +50,35 @@ static def getDate() {
     Date now = new Date()
     SimpleDateFormat yearMonthDateHourMin = new SimpleDateFormat("yyyyMMddHHmm")
     return yearMonthDateHourMin.format(now)
+}
+
+def parseComponentsFile(componentsFile, componentType) {
+    def componentConfigs = [:]
+    componentsJson = readFile componentsFile
+    def object = new JsonSlurperClassic().parseText componentsJson
+    def components = object['components']
+
+    for (i = 0; i < components.size(); i++) {
+        def index = i
+        if (components[index]['type'] == componentType) {
+            def config = components[index]
+            def componentName = config['name']
+            componentConfigs[componentName] = [:]
+            componentConfigs[componentName]['gitHubOrg'] = params.gitHubOrg ?: config['org']
+            componentConfigs[componentName]['repoName'] = config['repoName'] ?: componentName
+            componentConfigs[componentName]['baseBranch'] = config['baseBranch'] ?: 'master'
+            componentConfigs[componentName]['repoDir'] = config['repoDir'] ?: ''
+            componentConfigs[componentName]['cookbook'] = config['cookbook'] ?: componentName
+            componentConfigs[componentName]['buildType'] = config['buildType'] ?: 'node'
+            componentConfigs[componentName]['distCmd'] = config['distCmd']
+            componentConfigs[componentName]['labels'] = config?.labels ?: {}
+            componentConfigs[componentName]['type'] = config['type']
+            componentConfigs[componentName]['buildJobName'] = config['buildJobName'] ?: "build_any_jenkinsfile"
+            componentConfigs[componentName]['gitUrl'] = "git@github.com:${componentConfigs[componentName]['gitHubOrg']}/${componentConfigs[componentName]['repoName']}.git"
+            componentConfigs[componentName]['gitHubUrl'] = "https://github.com/${componentConfigs[componentName]['gitHubOrg']}/${componentConfigs[componentName]['repoName']}"
+        }
+    }
+    return componentConfigs
 }
 
 @NonCPS
