+++ vars/getComponentConfigs.groovy
@@ -1,16 +1,20 @@
 #!/usr/bin/groovy
 
-import groovy.json.JsonSlurperClassic
+import org.feedhenry.Utils
 
-def call(componentConfigs, configGitRepo, configGitRef, tryMasterOnFail = false) {
+def call(configGitRepo, configGitRef, tryMasterOnFail = false) {
+    def componentConfigs = [:]
     node {
         step([$class: 'WsCleanup'])
+
+        def utils = new Utils()
 
         try {
             checkoutGitRepo {
                 repoUrl = configGitRepo
                 branch = configGitRef
                 targetDir = 'product_releases'
+                shallow = true
             }
         } catch (Exception e) {
             if (tryMasterOnFail) {
@@ -18,6 +22,7 @@ def call(componentConfigs, configGitRepo, configGitRef, tryMasterOnFail = false)
                     repoUrl = configGitRepo
                     branch = 'master'
                     targetDir = 'product_releases'
+                    shallow = true
                 }
             } else {
                 throw e
@@ -25,30 +30,8 @@ def call(componentConfigs, configGitRepo, configGitRef, tryMasterOnFail = false)
         }
 
         dir('product_releases') {
-            componentsJson = readFile 'COMPONENTS.json'
-            def object = new JsonSlurperClassic().parseText componentsJson
-            def components = object['components']
-
-            for (i = 0; i < components.size(); i++) {
-                def index = i
-                if (components[index]['type'] == 'platform') {
-                    def config = components[index]
-                    def componentName = config['name']
-                    componentConfigs[componentName] = [:]
-                    componentConfigs[componentName]['gitHubOrg'] = params.gitHubOrg ?: config['org']
-                    componentConfigs[componentName]['repoName'] = config['repoName'] ?: componentName
-                    componentConfigs[componentName]['baseBranch'] = config['baseBranch'] ?: 'master'
-                    componentConfigs[componentName]['repoDir'] = config['repoDir'] ?: ''
-                    componentConfigs[componentName]['cookbook'] = config['cookbook'] ?: componentName
-                    componentConfigs[componentName]['buildType'] = config['buildType'] ?: 'node'
-                    componentConfigs[componentName]['distCmd'] = config['distCmd']
-                    componentConfigs[componentName]['labels'] = config?.labels ?: {}
-                    componentConfigs[componentName]['type'] = config['type']
-                    componentConfigs[componentName]['buildJobName'] = config['buildJobName'] ?: "build_any_jenkinsfile"
-                    componentConfigs[componentName]['gitUrl'] = "git@github.com:${componentConfigs[componentName]['gitHubOrg']}/${componentConfigs[componentName]['repoName']}.git"
-                    componentConfigs[componentName]['gitHubUrl'] = "https://github.com/${componentConfigs[componentName]['gitHubOrg']}/${componentConfigs[componentName]['repoName']}"
-                }
-            }
+            componentConfigs = utils.parseComponentsFile('COMPONENTS.json', 'platform')
         }
     }
+    return componentConfigs
 }
